XW-5041 | Making search logic work in search modal
