+++ app/templates/components/wds-global-navigation.hbs
@@ -38,6 +38,7 @@
 				navigationModel=model.main_navigation
 				isOpen=isSearchModalOpen
 				openModal=(action 'openModal')
+				activateSearch=(action 'activateSearch')
 				onSearchSuggestionChosen=(action 'onSearchSuggestionChosen')
 				goToSearchResults=(action 'goToSearchResults')
 			}}
