+++ app/templates/components/wds-global-navigation/search-modal.hbs
@@ -5,32 +5,13 @@
 {{#if isOpen}}
 	<div class="wds-global-navigation__modal wds-search-modal">
 		<div class="wds-search-modal__form-and-links-wrapper">
-			<form action="/search" method="POST" class="wds-search-modal__form">
-				{{wds-icon name='magnifying-glass' class='wds-search-modal__search-icon'}}
-				{{input
-					value=query
-					class='wds-search-modal__input'
-					name=model.results.param-name
-					placeholder=(i18n model.placeholder-active.key ns='design-system' sitename=model.placeholder-active.params.sitename.value)
-					type='search'
-					autocomplete='off'
-					input=(action 'onQueryChanged')
-				}}
-
-				{{#wds-list linked=true ellipsis=true}}
-					{{#each suggestions as |suggestion index|}}
-						<li class="{{if (equal index selectedSuggestionIndex) 'wds-is-selected'}}"
-							onmousedown={{action 'onSearchSuggestionChosen' suggestion}} onmouseenter={{action
-							'selectItem' index}}>
-							<a href="{{suggestion.uri}}"
-							   class="wds-global-navigation__dropdown-link"
-							   data-tracking-label="{{model.suggestions.tracking_label}}">
-								{{{suggestion.text}}}
-							</a>
-						</li>
-					{{/each}}
-				{{/wds-list}}
-			</form>
+			{{wds-global-navigation/search
+				model=model
+				activateSearch=activateSearch
+				deactivateSearch=deactivateSearch
+				onSearchSuggestionChosen=onSearchSuggestionChosen
+				goToSearchResults=goToSearchResults
+			}}
 			{{wds-global-navigation/main-navigation model=navigationModel}}
 		</div>
 	</div>
